+++ src/cvodes/cvodes_ml.h
@@ -60,8 +60,6 @@ void sunml_cvodes_check_flag(const char *call, int flag, void *cvode_mem);
 
 value sunml_cvodes_make_jac_arg(realtype t, N_Vector y, N_Vector yb,
 			  N_Vector fyb, value tmp);
-void sunml_cvodes_wrap_to_nvector_table(int n, value vy, N_Vector *y);
-
 
 // NB: overrides CHECK_FLAG macro in cvode_ml.h
 #define SCHECK_FLAG(call, flag) if (flag != CV_SUCCESS) \